# fix/chat: Improve performance in streaming responses in long chats