@dominiccooney 2024-10-11
It looks like this breaks @-mentions with line ranges. I changed the behavior of the hydrate function a bit, so I will look into that... I may need to update a caller.