@dominiccooney 2024-10-11
This is because the lazy Proxies are not cloneable, so if they're posted back then they cause an exception.