@dominiccooney 2024-10-11
Expect CI to cycle green, please take a look.

If this is too complicated, another simpler solution is to make onMessage take a fast filter and a slow callback. It makes like tedious for callers because they need to duplicate the set of objects they handle in their callback's `switch (message.type) ...`, and it might be slower for big messages which we only use subsets of. But with some API craft we can probably make something palatable, if this is too complex.